+++ src/components/Filters/FilterFile/UploadFile/DocumentsFieldArray.js
@@ -2,6 +2,7 @@ import PropTypes from 'prop-types';
 import { isEmpty } from 'lodash';
 import { injectIntl, FormattedMessage } from 'react-intl';
 import { Field } from 'react-final-form';
+import { useFieldArray } from 'react-final-form-arrays';
 
 import {
   Button,
@@ -10,9 +11,12 @@ import {
   Row,
   TextField,
 } from '@folio/stripes/components';
-import { EditCard } from '@folio/stripes-erm-components';
-import { useKiwtFieldArray } from '@k-int/stripes-kint-components';
 
+import EditCard from './EditCard/EditCard';
+import {
+  onAddField,
+  onDeleteField,
+} from './EditCard/editcard-util';
 import FileUploaderField from './FileUploaderField';
 import Required from '../../../DisplayUtils/Validate';
 
@@ -21,7 +25,7 @@ const DocumentsFieldArray = ({
   fields: { name },
   onUploadFile,
 }) => {
-  const { items, onAddField, onDeleteField } = useKiwtFieldArray(name);
+  const { fields } = useFieldArray(name);
 
   const renderFileUpload = (doc, i) => {
     if (isEmpty(doc.fileId)) {
@@ -59,7 +63,7 @@ const DocumentsFieldArray = ({
   };
 
   const renderDocs = () => {
-    return items.map((doc, i) => (
+    return fields.map((doc, i) => (
       <EditCard
         data-test-filter-file
         deletebuttonarialabel={`delete filter file ${name}`}
@@ -68,8 +72,8 @@ const DocumentsFieldArray = ({
           'data-test-delete-filter-file-button': true
         }}
         header={<FormattedMessage id="ui-finc-select.filter.file.label" values={{ number: i + 1 }} />}
-        key={i}
-        onDelete={() => onDeleteField(i, doc)}
+        key={`${name}[${i}]`}
+        onDelete={() => onDeleteField(fields, i, doc)}
       >
         <Row>
           <Col xs={12} md={onUploadFile ? 6 : 12}>
@@ -118,12 +122,12 @@ const DocumentsFieldArray = ({
   return (
     <div data-test-filter-file-card>
       <div>
-        { items.length ? renderDocs() : renderEmpty() }
+        { fields.length ? renderDocs() : renderEmpty() }
       </div>
       <Button
         data-test-filter-file-card-add-button
         id="add-filter-file-btn"
-        onClick={() => onAddField()}
+        onClick={() => onAddField(fields)}
       >
         <FormattedMessage id="ui-finc-select.filter.file.addFile" />
       </Button>